refactor(seller_leads): move seller leads out of interests
