+++ apps/re/lib/interests/policy.ex
@@ -6,8 +6,6 @@ defmodule Re.Interests.Policy do
 
   def authorize(_, %User{role: "admin"}, _), do: :ok
 
-  def authorize(:request_price_suggestion, %User{}, _), do: :ok
-
   def authorize(_, nil, _), do: {:error, :unauthorized}
 
   def authorize(_, _, _), do: {:error, :forbidden}
