+++ apps/re/lib/seller_leads/notify_when_covered.ex
@@ -1,4 +1,4 @@
-defmodule Re.Interests.NotifyWhenCovered do
+defmodule Re.SellerLeads.NotifyWhenCovered do
   @moduledoc """
   Schema for storing interest in non-covered regions
   """
