+++ apps/re/lib/seller_leads/policy.ex
@@ -8,6 +8,7 @@ defmodule Re.SellerLeads.Policy do
   def authorize(_, %User{role: "admin"}, _), do: :ok
   def authorize(:create_seller_lead, %User{role: "user"}, _), do: :ok
   def authorize(:create_broker_seller_lead, %User{role: "user"}, _), do: :ok
+  def authorize(:create_price_suggestion, %User{role: "user"}, _), do: :ok
   def authorize(_, nil, _), do: {:error, :unauthorized}
 
   def authorize(_, _, _), do: {:error, :forbidden}
