+++ apps/re_web/test/graphql/seller_leads/price_suggestions/query_test.exs
@@ -1,4 +1,4 @@
-defmodule ReWeb.GraphQL.Interests.PriceSuggestions.QueryTest do
+defmodule ReWeb.GraphQL.SellerLeads.PriceSuggestions.QueryTest do
   use ReWeb.ConnCase
 
   alias ReWeb.AbsintheHelpers
